Fix issue 5, updating Search and ForYou page
